Added support for SPM
