+++ ios/FluentIcons.podspec
@@ -28,8 +28,8 @@ FluentIcons
 
   s.subspec 'Code' do |sp|
     sp.source_files = 'ios/FluentIcons/Classes/*.swift'
-    sp.ios.source_files   = 'ios/FluentIcons/Classes/ios/*.swift'
-    sp.osx.source_files   = 'ios/FluentIcons/Classes/mac/*.swift'
+    sp.ios.source_files   = 'ios/FluentIcons/Classes/*.swift'
+    sp.osx.source_files   = 'ios/FluentIcons/Classes/*.swift'
   end
 
   s.subspec 'Assets' do |sp|
