+++ ios/FluentIcons/Classes/Image+FluentIcon.swift
@@ -3,6 +3,29 @@
 //  Licensed under the MIT license.
 //
 
+#if os(macOS)
+import AppKit
+
+private class FluentIconsBundleCheck {}
+
+public extension NSImage {
+  @objc static let fluentIconBundle = Bundle(for: FluentIconsBundleCheck.self)
+
+  @objc static func fluentIcon(_ fluent: FluentIcon) -> NSImage {
+    // Force unwrap here because the resource strings
+    // are generated so we can be confident that the image
+    // exits at runtime.
+    #if SWIFT_PACKAGE
+    return NSImage.fluentIconBundle.image(forResource: NSImage.Name(fluent.resourceString))!
+    #else
+    return NSImage.fluentIconBundle.image(forResource: NSImage.Name(fluent.resourceString))!
+    #endif
+  }
+}
+
+#endif
+
+#if os(iOS)
 import UIKit
 
 private class FluentIconsBundleCheck {}
@@ -14,7 +37,11 @@ public extension UIImage {
     // Force unwrap here because the resource strings
     // are generated so we can be confident that the image
     // exits at runtime.
+    #if SWIFT_PACKAGE
+    self.init(named: fluent.resourceString, in: Bundle.module, compatibleWith: nil)!
+    #else
     self.init(named: fluent.resourceString, in: UIImage.fluentIconBundle, compatibleWith: nil)!
+    #endif
   }
 }
 
@@ -29,3 +56,4 @@ public extension UIImageView {
     self.tintColor = tintColor
   }
 }
+#endif
